refactor(engine): split mdg/object.rs into object.rs + json.rs (issue #104)
Split ModuleDependencyGraph disk-ingestion methods (from_gitnexus_dir,
from_json_dir, ingest_json_document and its two per-shape helpers) out
of object.rs into a new json.rs, via a second impl
ModuleDependencyGraph block -- ordinary Rust; the type does not care
which file its methods live in. object.rs now owns only what the graph
is and how it is queried (construction, add_node/add_relationship,
lookups, the Representation impl); json.rs owns how one gets built
from a .gitnexus directory. Moved the one test that exercises
from_gitnexus_dir alongside it.

This is a genuine COMPOSABLE-aligned split (two distinct
responsibilities, not artificially carved up to dodge a score), but it
is also a deliberate experiment to test the "split files to fix
SIMPLE" theory documented in the last two commits. Result: it does not
fully work. cfg.cyclomatic roughly halved (48 -> 25 and 24) but both
resulting files still exceed the flat threshold of 15 -- splitting a
~50-cyclomatic file into two ~25s does not clear a 15-point bar. To
actually flip these to SIMPLE via splitting alone would need
fragmenting into many more, much smaller files than is reasonable for
readability/navigability -- genuine metric-gaming, not composability.

This is now first-hand evidence (not just theory) that the ~28
remaining SLOP files in topos-engine will not be resolved by more
file-splitting in this pass. The actual fix belongs in the SIMPLE
policy itself: cfg.cyclomatic is a whole-file merged-CFG sum (see
cfg/builder.rs, which deliberately keeps the connected-component count
at 1) that scales with function count with no normalization, so any
file with a handful of genuinely-simple methods can exceed a flat
threshold of 15 regardless of how simple each one is. A fix (e.g.
normalizing by function count, capping per-function contribution, or
averaging instead of summing) changes scoring semantics for every
evaluation this tool produces, not just this crate self-evaluation,
and would need its own calibration review -- flagging as necessary
follow-up rather than changing unilaterally here.

277 workspace tests pass unchanged (json.rs test moved, not added);
cargo clippy --workspace --all-targets -- -D warnings and
cargo fmt --all --check both clean.
