+++ topos/engine/src/graphs/mdg/mod.rs
@@ -1,5 +1,6 @@
 //! The Module Dependency Graph — the inter-module (import/call/inherit)
 //! view parsed from GitNexus output, feeding the COMPOSABLE generator.
 
+mod json;
 pub mod models;
 pub mod object;
